+++ Civi/ActionProvider/Action/Relationship/CreateRelationshipWithTypeParameter.php
@@ -7,6 +7,7 @@
 namespace Civi\ActionProvider\Action\Relationship;
 
 use \Civi\ActionProvider\Action\AbstractAction;
+use Civi\ActionProvider\ConfigContainer;
 use \Civi\ActionProvider\Parameter\ParameterBagInterface;
 use \Civi\ActionProvider\Parameter\SpecificationBag;
 use \Civi\ActionProvider\Parameter\Specification;
@@ -69,14 +70,11 @@ class CreateRelationshipWithTypeParameter extends AbstractAction {
       new Specification('description', 'String', E::ts('Description'), false),
     ));
 
-    $customGroups = civicrm_api3('CustomGroup', 'get', array('extends' => 'Relationship', 'is_active' => 1, 'options' => array('limit' => 0)));
-    foreach($customGroups['values'] as $customGroup) {
-      $customFields = civicrm_api3('CustomField', 'get', array('custom_group_id' => $customGroup['id'], 'is_active' => 1, 'options' => array('limit' => 0)));
-      foreach($customFields['values'] as $customField) {
-        $spec = CustomField::getSpecFromCustomField($customField, $customGroup['title'].': ', false);
-        if ($spec) {
-          $specs->addSpecification($spec);
-        }
+    $config = ConfigContainer::getInstance();
+    $customGroups = $config->getCustomGroupsForEntity('Relationship');
+    foreach ($customGroups as $customGroup) {
+      if (!empty($customGroup['is_active'])) {
+        $specs->addSpecification(CustomField::getSpecForCustomGroup($customGroup['id'], $customGroup['name'], $customGroup['title']));
       }
     }
     return $specs;
@@ -132,7 +130,7 @@ class CreateRelationshipWithTypeParameter extends AbstractAction {
         continue;
       }
       if ($parameters->doesParameterExists($spec->getName())) {
-        list($customFieldID, $customValueID) = \CRM_Core_BAO_CustomField::getKeyID($spec->getApiFieldName(), TRUE);
+        [$customFieldID, $customValueID] = \CRM_Core_BAO_CustomField::getKeyID($spec->getApiFieldName(), TRUE);
         $value = $parameters->getParameter($spec->getName());
         if (is_array($value)) {
           $value = \CRM_Core_DAO::VALUE_SEPARATOR . implode(\CRM_Core_DAO::VALUE_SEPARATOR, $value) . \CRM_Core_DAO::VALUE_SEPARATOR;