+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version 1.56 (not yet released)
 ------------
 
+* Improved performance of various actions by using the config container which caches the api calls of custom groups and custom fields.
+
 Version 1.55
 ------------